Render commit relative-time client-side with accurate elapsed hours
