+++ src/types/github.ts
@@ -4,14 +4,4 @@ export interface CommitInfo {
   date: string;
   additions: number;
   deletions: number;
-  relativeTime: string;
 }
-
-const FALLBACK: CommitInfo = {
-  sha: "unknown",
-  message: "",
-  date: "",
-  additions: 0,
-  deletions: 0,
-  relativeTime: "unavailable",
-};
